refactor: PointsGenerator util패키지로 이동, PointerGenerator에서 domain 패키지에 대한 의존성 제거
