--- src/main/java/nextstep/ladder/domain/PointsGenerator.java
@@ -1,7 +1,0 @@
-package nextstep.ladder.domain;
-
-import java.util.List;
-
-public interface PointsGenerator {
-    List<Boolean> getPoints(NumberOfParticipants numberOfParticipants);
-}
