+++ src/main/java/nextstep/ladder/util/PointsGenerator.java
@@ -1,0 +1,7 @@
+package nextstep.ladder.util;
+
+import java.util.List;
+
+public interface PointsGenerator {
+    List<Boolean> getPoints(int pointsSize);
+}
